combine separate python files into one application
